Add more test results

+++ fun-bomb-cs/Program.cs
@@ -7,12 +7,12 @@ var http_client = new HttpClient();
 
 var isDoneLocker = new object();
 var isDone = false;
-var tasks = new List<Task<int>>();
+var tasks = new List<Task<long>>();
 
 var timer = Stopwatch.StartNew();
 for (int i = 0; i < connections; i++) {
     tasks.Add(Task.Run(async () => {
-        var count = 0;
+        var count = 0L;
         while (true) {
             lock (isDoneLocker) {
                 if (isDone) return count;
@@ -36,5 +36,3 @@ var successCount = tasks.Select(t => t.Result).Sum();
 var rps = successCount * 1000 / timeCost;
 
 Console.WriteLine($"Actual time {timeCost / 1000.0,0:N2}s, RPS {rps}/s.");
-
-
